refactor: create global.h absorbing all global headers, convert windef.h + c16string.h + always.h to forwarding headers
- Add original/global.h with 17 sections absorbing windef.h, c16string.h,
  clangcompat.h, floattypes.h, and always.h content
- Convert compat/windef.h and compat/c16string.h to minimal forwarding headers
- Convert Code/wwlib/always.h and Code/Scripts/always.h to forwarding headers
- Delete compat/clangcompat.h and compat/floattypes.h (content absorbed into global.h)
- Add ${CMAKE_SOURCE_DIR} to include_directories so global.h is found at source root
- Add ULONG, ULONG_PTR to global.h (needed by oaidl.h and winbase.h)
- Sync worktree with main branch wchar_t→char16_t changes across wwui, compat, and other modules
- Add winfile.cpp to Commando build (implements FindFirstFile/FindNextFile via POSIX)

+++ original/Code/Combat/control.cpp
@@ -19,7 +19,7 @@ ControlClass & ControlClass::operator = (const ControlClass & src)
 {
 	OneTimeBooleanBits		= src.OneTimeBooleanBits;
 	ContinuousBooleanBits	= src.ContinuousBooleanBits;
-	memcpy( AnalogValues, src.AnalogValues, sizeof(AnalogValues) ); 
+	memcpy( AnalogValues, src.AnalogValues, sizeof(AnalogValues) );
 	return *this;
 }
 
@@ -65,8 +65,8 @@ bool	ControlClass::Load( ChunkLoadClass &cload )
 						READ_MICRO_CHUNK( cload, MICROCHUNKID_CONTINUOUS_BOOL, ContinuousBooleanBits );
 						READ_MICRO_CHUNK( cload, MICROCHUNKID_PENDING_ONE_TIME_BOOL, PendingOneTimeBooleanBits );
 						READ_MICRO_CHUNK( cload, MICROCHUNKID_PENDING_CONTINUOUS_BOOL, PendingContinuousBooleanBits );
-						case MICROCHUNKID_ANALOG:	
-							cload.Read( &AnalogValues[0],sizeof(AnalogValues) ); 
+						case MICROCHUNKID_ANALOG:
+							cload.Read( &AnalogValues[0],sizeof(AnalogValues) );
 							break;
 
 						default:
@@ -89,19 +89,19 @@ bool	ControlClass::Load( ChunkLoadClass &cload )
 }
 
 //JITTER
-void	ControlClass::Clear_Boolean( void )	
-{ 
+void	ControlClass::Clear_Boolean( void )
+{
 
 	OneTimeBooleanBits	= 0;
 	ContinuousBooleanBits	= 0;
 }
 
-void	ControlClass::Clear_Control( void )	
-{ 
+void	ControlClass::Clear_Control( void )
+{
 
 	OneTimeBooleanBits	= 0;
 	ContinuousBooleanBits	= 0;
-	memset( AnalogValues, 0, sizeof(AnalogValues) ); 
+	memset( AnalogValues, 0, sizeof(AnalogValues) );
 }
 
 enum {
@@ -117,9 +117,9 @@ enum {
 
 //-----------------------------------------------------------------------------
 void	ControlClass::Import_Cs( BitStreamClass & packet )
-{ 
+{
 	// Or in the new one time bits
-	ULONG otb_bits = packet.Get(otb_bits, BITPACK_ONE_TIME_BOOLEAN_BITS);
+	uint32_t otb_bits = packet.Get(otb_bits, BITPACK_ONE_TIME_BOOLEAN_BITS);
 	OneTimeBooleanBits |= otb_bits;
 
 	packet.Get(ContinuousBooleanBits, BITPACK_CONTINUOUS_BOOLEAN_BITS);
@@ -167,7 +167,7 @@ void	ControlClass::Import_Cs( BitStreamClass & packet )
 	for (int i = 0; i < ANALOG_CONTROL_COUNT; i++) {
 		if (fabs(AnalogValues[i]) < 0.005) {
 			//
-			// The code elsewhere has no tolerance for precision error on this 
+			// The code elsewhere has no tolerance for precision error on this
 			// value around zero...
 			//
 			AnalogValues[i] = 0;
@@ -176,7 +176,7 @@ void	ControlClass::Import_Cs( BitStreamClass & packet )
 }
 
 //-----------------------------------------------------------------------------
-void ControlClass::Export_Cs( BitStreamClass & packet ) 
+void ControlClass::Export_Cs( BitStreamClass & packet )
 {
 //	packet.Add(OneTimeBooleanBits,		BITPACK_ONE_TIME_BOOLEAN_BITS);
 	packet.Add(PendingOneTimeBooleanBits,		BITPACK_ONE_TIME_BOOLEAN_BITS);
@@ -223,7 +223,7 @@ void ControlClass::Export_Cs( BitStreamClass & packet )
 
 //-----------------------------------------------------------------------------
 void ControlClass::Import_Sc(BitStreamClass & packet)
-{ 
+{
 	packet.Get(ContinuousBooleanBits, BITPACK_CONTINUOUS_BOOLEAN_BITS);
 
 #if 01
@@ -268,7 +268,7 @@ void ControlClass::Import_Sc(BitStreamClass & packet)
 	for (int i = 0; i < ANALOG_CONTROL_COUNT; i++) {
 		if (fabs(AnalogValues[i]) < 0.005) {
 			//
-			// The code elsewhere has no tolerance for precision error on this 
+			// The code elsewhere has no tolerance for precision error on this
 			// value around zero...
 			//
 			AnalogValues[i] = 0;
@@ -277,7 +277,7 @@ void ControlClass::Import_Sc(BitStreamClass & packet)
 }
 
 //-----------------------------------------------------------------------------
-void ControlClass::Export_Sc(BitStreamClass & packet) 
+void ControlClass::Export_Sc(BitStreamClass & packet)
 {
 	packet.Add(ContinuousBooleanBits,	BITPACK_CONTINUOUS_BOOLEAN_BITS);
 
@@ -319,7 +319,7 @@ void ControlClass::Export_Sc(BitStreamClass & packet)
 
 //-----------------------------------------------------------------------------
 /*
-float	ControlClass::Get_Clamp(AnalogControl control) 
+float	ControlClass::Get_Clamp(AnalogControl control)
 {
    //
    // This will definitely need improving...
@@ -354,12 +354,12 @@ float	ControlClass::Get_Clamp(AnalogControl control)
 */
 
 /*
-BYTE ControlClass::Scale_Analog(float clamp, float unscaled) 
+BYTE ControlClass::Scale_Analog(float clamp, float unscaled)
 {
    return (BYTE) (255 * WWMath::Clamp((unscaled + clamp) / (2 * clamp)));
 }
 
-float ControlClass::Unscale_Analog(float clamp, BYTE scaled) 
+float ControlClass::Unscale_Analog(float clamp, BYTE scaled)
 {
    float unscaled;
    if (scaled == 127) { // scaling perturbs zero, special-case it
@@ -372,32 +372,32 @@ float ControlClass::Unscale_Analog(float clamp, BYTE scaled)
 */
 
 void	ControlClass::Set_Boolean( BooleanControl bcontrol, bool state )
-{ 
+{
 	int control = bcontrol;
 	if ( control >= BOOLEAN_CONTINUOUS_FIRST ) {
 		control -= BOOLEAN_CONTINUOUS_FIRST;
-		if (state) 	ContinuousBooleanBits |= (1 << control); 
-		else 		  	ContinuousBooleanBits &= ~(1 << control); 
+		if (state) 	ContinuousBooleanBits |= (1 << control);
+		else 		  	ContinuousBooleanBits &= ~(1 << control);
 
 		PendingContinuousBooleanBits |= ContinuousBooleanBits;
 	} else {
 		control -= BOOLEAN_ONE_TIME_FIRST;
-		if (state) 	OneTimeBooleanBits |= (1 << control); 
-		else 		  	OneTimeBooleanBits &= ~(1 << control); 
+		if (state) 	OneTimeBooleanBits |= (1 << control);
+		else 		  	OneTimeBooleanBits &= ~(1 << control);
 
 		PendingOneTimeBooleanBits |= OneTimeBooleanBits;
 	}
 }
 
 bool	ControlClass::Get_Boolean( BooleanControl bcontrol )
-{ 
+{
 	int control = bcontrol;
 	if ( control >= BOOLEAN_CONTINUOUS_FIRST ) {
 		control -= BOOLEAN_CONTINUOUS_FIRST;
-		return ( ContinuousBooleanBits & (1 << control ) ) ? true : false; 
+		return ( ContinuousBooleanBits & (1 << control ) ) ? true : false;
 	} else {
 		control -= BOOLEAN_ONE_TIME_FIRST;
-		return ( OneTimeBooleanBits & (1 << control ) ) ? true : false; 
+		return ( OneTimeBooleanBits & (1 << control ) ) ? true : false;
 	}
 }
 
@@ -410,7 +410,7 @@ void ControlClass::Set_Precision(void)
 	cEncoderList::Set_Precision(BITPACK_CONTROL_MOVES_SC,				CONTROL_MOVE_DOWN+1); // 6
 
 	//
-	// TSS: I think that analog values are within this range, except for 
+	// TSS: I think that analog values are within this range, except for
 	// DEBUG_RAPID_MOVE...
 	//
 	cEncoderList::Set_Precision(BITPACK_ANALOG_VALUES,	-1.0, 1.0, 0.01);
