+++ original/Code/Commando/CDKeyAuth.cpp
@@ -2,7 +2,7 @@
 
 void CCDKeyAuth::GetSerialNum(StringClass &serial) { serial = ""; }
 void CCDKeyAuth::DisconnectUser(int localid) {}
-void CCDKeyAuth::AuthenticateUser(int localid, ULONG ip, char *challenge, char *authstring) {}
+void CCDKeyAuth::AuthenticateUser(int localid, uint32_t ip, char *challenge, char *authstring) {}
 char *CCDKeyAuth::GenChallenge(int nchars) { return nullptr; }
 void CCDKeyAuth::AuthSerial(const char *challenge, StringClass &resp) { resp = ""; }
 void CCDKeyAuth::auth_callback(int localid, int authenticated, char *errmsg, void *instance) {}
