+++ original/Code/Commando/CDKeyAuth.h
@@ -20,7 +20,7 @@ protected:
 public:
 	static void GetSerialNum(StringClass &serial);
 	static void DisconnectUser(int localid);
-	static void AuthenticateUser(int localid, ULONG ip, char *challenge, char *authstring);
+	static void AuthenticateUser(int localid, uint32_t ip, char *challenge, char *authstring);
 	static char *GenChallenge(int nchars);
 	static void AuthSerial(const char *challenge, StringClass &resp);
