+++ original/Code/Commando/useroptions.cpp
@@ -94,7 +94,7 @@ bool cUserOptions::Parse_Command_Line(LPCSTR command)
 
 		// Look for ip override.
 		if (strstr(cmd, "IP=")) {
-			extern ULONG g_ip_override;
+			extern uint32_t g_ip_override;
 			g_ip_override = ::inet_addr(strstr(cmd, "IP=") + 3);
 			continue;
 		}
@@ -262,10 +262,10 @@ void cUserOptions::Set_Bandwidth_Type(BANDWIDTH_TYPE_ENUM bandwidth_type)
 
 	if (bandwidth_type != BANDWIDTH_CUSTOM) {
 		if (bandwidth_type == BANDWIDTH_AUTO && BandwidthCheckerClass::Got_Bandwidth()) {
-			ULONG bps = BandwidthCheckerClass::Get_Upstream_Bandwidth();
+			uint32_t bps = BandwidthCheckerClass::Get_Upstream_Bandwidth();
 			BandwidthBps.Set(bps);
 		} else {
-			ULONG bps = cBandwidth::Get_Bandwidth_Bps_From_Type(bandwidth_type);
+			uint32_t bps = cBandwidth::Get_Bandwidth_Bps_From_Type(bandwidth_type);
 			BandwidthBps.Set(bps);
 		}
 	}
@@ -316,7 +316,7 @@ cRegistryBool cUserOptions::GameListFilterShowOnlyGamesIRankFor(	APPLICATION_SUB
 		value = ::strstr(cmd, param);
 		if (value != NULL) {
 			value += ::strlen(param);
-			ULONG ip = ::inet_addr(value);
+			uint32_t ip = ::inet_addr(value);
 			cGameSpyAdmin::Set_Game_Host_Ip(ip);
 			cGameSpyAdmin::Set_Is_Launch_From_Gamespy_Requested(true);
 			continue;
