+++ original/Code/ww3d2/agg_def.h
@@ -49,23 +49,23 @@ class AggregateDefClass
 		///////////////////////////////////////////////////////////
 		//
 		//	Public constructors/destructors
-		//		
+		//
 		AggregateDefClass (void);
 		AggregateDefClass (RenderObjClass &base_model);
 		AggregateDefClass (const AggregateDefClass &src);
 		virtual ~AggregateDefClass (void);
 
-		
+
 		///////////////////////////////////////////////////////////
 		//
 		//	Public operators
-		//		
+		//
 		const AggregateDefClass &operator= (const AggregateDefClass &src);
 
 		///////////////////////////////////////////////////////////
 		//
 		//	Public methods
-		//		
+		//
 		virtual WW3DErrorType	Load_W3D (ChunkLoadClass &chunk_load);
 		virtual WW3DErrorType	Save_W3D (ChunkSaveClass &chunk_save);
 		const char *				Get_Name (void) const					{ return m_pName; }
@@ -76,7 +76,7 @@ class AggregateDefClass
 		//
 		//	Public accessors
 		//
-		ULONG							Class_ID (void) const					{ return m_MiscInfo.OriginalClassID; }
+		uint32_t							Class_ID (void) const					{ return m_MiscInfo.OriginalClassID; }
 
 		//
 		//	Initialization
@@ -97,13 +97,13 @@ class AggregateDefClass
 			bool operator == (_TEXTURE_INFO &src) { return false; }
 			bool operator != (_TEXTURE_INFO &src) { return true; }
 		} TEXTURE_INFO;
-		
+
 
 		///////////////////////////////////////////////////////////
 		//
 		//	Protected methods
 		//
-		
+
 		//
 		//	Loading methods
 		//
@@ -158,17 +158,17 @@ class AggregateDefClass
 //
 //	AggregatePrototypeClass
 //
-class AggregatePrototypeClass : public PrototypeClass 
+class AggregatePrototypeClass : public PrototypeClass
 {
 	public:
 
 		///////////////////////////////////////////////////////////
 		//
 		//	Public constructors/destructors
-		//		
+		//
 		AggregatePrototypeClass (AggregateDefClass *pdef)		{ m_pDefinition = pdef; }
 		virtual ~AggregatePrototypeClass (void)					{ delete m_pDefinition; }
-		
+
 		///////////////////////////////////////////////////////////
 		//
 		//	Public methods
