+++ original/Code/ww3d2/shader.cpp
@@ -226,7 +226,7 @@ void ShaderClass::Init_From_Material3(const W3dMaterial3Struct & mat3)
  * WARNINGS:                                                                                   *
  *                                                                                             *
  * HISTORY:																												  *
- *		05/02/00    IML : Created.																					  *	
+ *		05/02/00    IML : Created.																					  *
  *=============================================================================================*/
 
 void ShaderClass::Enable_Fog (const char *source)
@@ -239,12 +239,12 @@ void ShaderClass::Enable_Fog (const char *source)
 				Set_Fog_Func (ShaderClass::FOG_WHITE);
 			} else {
 				Report_Unable_To_Fog (source);
-			}	
+			}
 			break;
 
 		case ShaderClass::SRCBLEND_ONE:
 			switch (Get_Dst_Blend_Func()) {
-				
+
 				case ShaderClass::DSTBLEND_ZERO:							// Opaque.
 					Set_Fog_Func (ShaderClass::FOG_ENABLE);
 					break;
@@ -265,7 +265,7 @@ void ShaderClass::Enable_Fog (const char *source)
 				Set_Fog_Func (ShaderClass::FOG_ENABLE);
 			} else {
 				Report_Unable_To_Fog (source);
-			}	
+			}
 			break;
 
 		case ShaderClass::SRCBLEND_ONE_MINUS_SRC_ALPHA:
@@ -273,13 +273,13 @@ void ShaderClass::Enable_Fog (const char *source)
 				Set_Fog_Func (ShaderClass::FOG_ENABLE);
 			} else {
 				Report_Unable_To_Fog (source);
-			}	
+			}
 			break;
 	}
 }
 
 /***********************************************************************************************
- * ShaderClass::Report_Unable_To_Fog --																		  *	
+ * ShaderClass::Report_Unable_To_Fog --																		  *
  *                                                                                             *
  * INPUT:                                                                                      *
  *                                                                                             *
@@ -288,7 +288,7 @@ void ShaderClass::Enable_Fog (const char *source)
  * WARNINGS:                                                                                   *
  *                                                                                             *
  * HISTORY:																												  *
- *		10/04/00    IML : Created.																					  *	
+ *		10/04/00    IML : Created.																					  *
  *=============================================================================================*/
 void ShaderClass::Report_Unable_To_Fog (const char *source)
 {
@@ -308,7 +308,7 @@ public:
 	bool		useAlpha;
 };
 
-const Blend srcBlendLUT[ShaderClass::SRCBLEND_MAX] = 
+const Blend srcBlendLUT[ShaderClass::SRCBLEND_MAX] =
 {
 	Blend(D3DBLEND_ZERO, false),
 	Blend(D3DBLEND_ONE, false),
@@ -316,7 +316,7 @@ const Blend srcBlendLUT[ShaderClass::SRCBLEND_MAX] =
  	Blend(D3DBLEND_INVSRCALPHA, true)
 };
 
-const Blend dstBlendLUT[ShaderClass::DSTBLEND_MAX] = 
+const Blend dstBlendLUT[ShaderClass::DSTBLEND_MAX] =
 {
 	Blend(D3DBLEND_ZERO, false),
 	Blend(D3DBLEND_ONE, false),
@@ -361,7 +361,7 @@ void ShaderClass::Apply()
 
 	if(diff & (ShaderClass::MASK_COLORMASK | ShaderClass::MASK_SRCBLEND | ShaderClass::MASK_DSTBLEND | ShaderClass::MASK_ALPHATEST))
 	{
-		ULONG planeMask = 0xffffff;
+		uint32_t planeMask = 0xffffff;
 
 		if(Get_Color_Mask() != ShaderClass::COLOR_WRITE_ENABLE)
 			planeMask = 0;
@@ -398,7 +398,7 @@ void ShaderClass::Apply()
 		if(Get_Alpha_Test() == ShaderClass::ALPHATEST_ENABLE)
 		{
 			unsigned char alphareference = 0x60;	// Alpha reference value that produces best results with mip-mapped textures.
-			
+
 			if(sf == D3DBLEND_INVSRCALPHA)
 			{
 				DX8Wrapper::Set_DX8_Render_State(D3DRS_ALPHAREF,0xff - alphareference);
@@ -427,14 +427,14 @@ void ShaderClass::Apply()
 
 			BOOL fm = FALSE;
 			D3DCOLOR fogColor = DX8Wrapper::Get_Fog_Color();
-			
+
 			switch(Get_Fog_Func())
 			{
 			case ShaderClass::FOG_ENABLE:
 				fm = TRUE;
 				break;
 			case ShaderClass::FOG_SCALE_FRAGMENT:
-				fogColor = 0;	
+				fogColor = 0;
 				fm = TRUE;
 				break;
 			case ShaderClass::FOG_WHITE:
@@ -456,7 +456,7 @@ void ShaderClass::Apply()
 		} else {
 			DX8Wrapper::Set_DX8_Render_State(D3DRS_FOGENABLE,FALSE);
 		}
-		
+
 		diff &= ~(ShaderClass::MASK_FOG);
 		if(!diff)
 			return;
@@ -493,7 +493,7 @@ void ShaderClass::Apply()
 				aArg2 = D3DTA_DIFFUSE;
 				break;
 			case ShaderClass::GRADIENT_ADD:
-				if(!(TextureOpCaps & D3DTEXOPCAPS_ADD))	
+				if(!(TextureOpCaps & D3DTEXOPCAPS_ADD))
 					cOp = D3DTOP_MODULATE;
 				else
 					cOp = D3DTOP_ADD;
@@ -797,7 +797,7 @@ void ShaderClass::Apply()
 	}
 
 	// Enable/disable alpha test
-	DX8Wrapper::Set_DX8_Render_State(D3DRS_ALPHATESTENABLE,BOOL(Get_Alpha_Test()));	
+	DX8Wrapper::Set_DX8_Render_State(D3DRS_ALPHATESTENABLE,BOOL(Get_Alpha_Test()));
 
 	// Enable/disable stencil test
 	// Not supported yet
